Restructured PDFoutline.py
